requirements.txt
Django==1.7.1
Jinja2==2.7.3
Markdown==2.5.2
MarkupSafe==0.23
Pillow==2.6.1
PyYAML==3.11
boto==2.35.1
certifi==14.05.14
coverage==3.7.1
django-appconf==0.6
django-countries==3.0.1
django-debug-toolbar==1.2.2
django-extensions==1.4.9
django-filter==0.9.1
django-haystack==2.3.1
django-nose==1.3
django-storages-redux==1.2.1
djangorestframework==3.0.3
elasticsearch==1.3.0
geopy==1.7.0
gnureadline==6.3.3
gunicorn==19.1.1
ipython==2.3.1
nose==1.3.4
polib==1.0.5
psycopg2==2.5.4
pyelasticsearch==0.7.1
pytz==2014.10
pyzmq==14.4.1
requests==2.5.1
simplejson==3.6.5
six==1.9.0
sqlparse==0.1.14
tornado==4.0.2
urllib3==1.10
